// backend/src/webhooks/dto/telegram-webhook.dto.ts
export class TelegramMessageDto {
  message_id: number;
  from: {
    id: number;
    is_bot: boolean;
    first_name: string;
    username?: string;
    language_code?: string;
  };
  chat: {
    id: number;
    first_name?: string;
    username?: string;
    type: string;
  };
  date: number;
  text?: string;
}

export class TelegramUpdateDto {
  update_id: number;
  message?: TelegramMessageDto;
}